Renaming noncontainer to sandwich
