+++ src/mpeg/mpegAudioFile.ts
@@ -1,6 +1,6 @@
 import MpegAudioFileSettings from "./mpegAudioFileSettings";
 import MpegAudioHeader from "./mpegAudioHeader";
-import NonContainerFile from "../nonContainer/nonContainerFile";
+import SandwichFile from "../sandwich/sandwichFile";
 import Properties from "../properties";
 import {CorruptFileError} from "../errors";
 import {File, ReadStyle} from "../file";
@@ -8,14 +8,14 @@ import {IFileAbstraction} from "../fileAbstraction";
 import {TagTypes} from "../tag";
 
 /**
- * This class extends {@link NonContainerFile} to provide tagging and properties support for
+ * This class extends {@link SandwichFile} to provide tagging and properties support for
  * MPEG-1, MPEG-2, and MPEG-2.5 non-containerized audio files.
  * @remarks A {@link Id3v1Tag} and {@link Id3v2Tag} will be added automatically to any file
  *     that doesn't contain one. This change does not affect the file until it is saved and can be
  *     reversed using the following method:
  *     `file.removeTags(file.tagTypes & ~file.tagTypesOnDisk);`
  */
-export default class MpegAudioFile extends NonContainerFile {
+export default class MpegAudioFile extends SandwichFile {
     private static readonly _defaultTagLocationMapping = new Map<TagTypes, () => boolean>([
         [TagTypes.Ape, () => MpegAudioFileSettings.preferApeTagAtFileEnd],
         [TagTypes.Id3v1, () => true],
