+++ src/mpeg/mpegContainerFile.ts
@@ -1,6 +1,6 @@
 import MpegAudioHeader from "./mpegAudioHeader";
 import MpegVideoHeader from "./mpegVideoHeader";
-import NonContainerFile from "../nonContainer/nonContainerFile";
+import SandwichFile from "../sandwich/sandwichFile";
 import Properties from "../properties";
 import {ByteVector} from "../byteVector";
 import {CorruptFileError, UnsupportedFormatError} from "../errors";
@@ -61,14 +61,14 @@ enum MpegFileMarker {
 }
 
 /**
- * This class extends {@link NonContainerFile} to provide tagging and properties support for
+ * This class extends {@link SandwichFile} to provide tagging and properties support for
  * MPEG-1, MPEG-2, and MPEG-2.5 containerized video files.
  * @remarks A {@link Id3v1Tag} and {@link Id3v2Tag} will be added automatically to any file that
  *     does not contain one. This change does not affect the file until it is saved and can be
  *     reversed using the following method:
  *     `file.removeTags(file.tagTypes & ~file.tagTypesOnDisk);`
  */
-export default class MpegContainerFile extends NonContainerFile {
+export default class MpegContainerFile extends SandwichFile {
     private static readonly _defaultTagLocationMapping = new Map<TagTypes, () => boolean>([
         [TagTypes.Ape, () => true],
         [TagTypes.Id3v1, () => true],
