+++ src/sandwich/endTag.ts
@@ -4,7 +4,7 @@ import Id3v2Settings from "../id3v2/id3v2Settings";
 import Id3v1Tag from "../id3v1/id3v1Tag";
 import Id3v2Tag from "../id3v2/id3v2Tag";
 import Id3v2TagFooter from "../id3v2/id3v2TagFooter";
-import TagParser from "../startEndTags/tagParsers";
+import TagParser from "./tagParsers";
 import {ApeTagFooter} from "../ape/apeTagFooter";
 import {ByteVector} from "../byteVector";
 import {CorruptFileError, UnsupportedFormatError} from "../errors";
