+++ src/sandwich/startTag.ts
@@ -2,7 +2,7 @@ import ApeTag from "../ape/apeTag";
 import CombinedTag from "../combinedTag";
 import Id3v2Tag from "../id3v2/id3v2Tag";
 import Id3v2Settings from "../id3v2/id3v2Settings";
-import TagParser from "../startEndTags/tagParsers";
+import TagParser from "./tagParsers";
 import {ApeTagFooter} from "../ape/apeTagFooter";
 import {ByteVector} from "../byteVector";
 import {CorruptFileError, UnsupportedFormatError} from "../errors";
